Experimental change removes requirement for iOS 10
This is work in progress, not yet tested. The core idea 
is to remove the @available check around the plugin 
and then do anything iOS 10 specific inside version 
specific if tests with casts. Because it is less readable and
testable than the previous version likely not going to merge
with master.