Port speech_to_text to Dart 2.12.
The upgrade to Dart 2.12 adds support for null safety.  This is a
breaking change, and removes support for older versions of Dart.

Update dependencies:
 - json_annotation 4.0.0
 - json_serializable 4.0.2
 - mockito 5.0.0
 - plugin_platform_interface 1.1.0-nullsafety.2

Stick to an older plugin_platform_interface for now because
that's what got merged into the interface PR.  We can roll both
when this is done.

Tested:
All tests pass.