Fix lint warning for control flow in finally
https://pub.dev/packages/speech_to_text/versions/4.1.0-nullsafety/score
has a warning and a lower score because of the return in the finally
block in web handling.  This currently
costs 10 points in the "Pass static analysis" score.

I think the change is worthwhile both from not having the package
penalized, but also because it's easy to see what the return state is
now from the variable name.

Tested:
Not at all.